refactor: extract service probe logic into shared module
